fix(tools): psv_dedup_partition の clippy / PR 472 レビュー対応
- clippy (unnecessary_cast): `stat.f_bavail as u64 * stat.f_frsize as u64`
  は Linux x86_64 で既に u64 同士。キャスト削除
- P1 (データ欠損リスク修正): --phase2-only で Phase 1 時の partition 数を
  temp dir から自動検出し、CLI --partitions と不一致なら検出値で上書きする。
  partition_NNNNN.bin の最大 index + 1 を使うので途中削除されても Phase 1
  時の N を復元可能。input/ref 間の partition 数不一致は Err。
- P2 (reference サイズ誤計上修正): ResourceEstimate に output_upper_bound_bytes
  を追加し、出力空き容量チェックは input 側のサイズのみを基準にする。
  reference は出力対象外なので除外。

+++ crates/tools/src/common/dedup.rs
@@ -149,7 +149,7 @@ pub fn get_disk_available(path: &Path) -> Option<u64> {
     if rc != 0 {
         return None;
     }
-    Some(stat.f_bavail as u64 * stat.f_frsize as u64)
+    Some(stat.f_bavail * stat.f_frsize)
 }
 
 /// バイト値を `%.1f GiB` 形式で整形する。
